Support require() relative to requiring file
Factorio allows requiring files relative to the file doing the require call.
For example given the scenario

modname_1.2.3/
`-- prototypes
    |-- a.lua
    `-- b.lua

both of the following require() calls are a valid way
to require file b from file a:

a) require("prototypes.b")
b) require("b")

This behaviour is different from default Lua, where require-calls must be
absolute to the base folder and where only case a) is valid.

This commit stores each file's directory in a stack as they are executed
and then tries to require the file both relative to the top of the
stacktrace and absolute to the root folder of the zip archive.